convert feed generation into broccoli plugin
